refactor: moved components used in home page to pages/home folder
