+++ sites/geohub/src/components/pages/home/MapHero.svelte
