+++ sites/geohub/src/components/pages/home/MapStyleCard.svelte
