+++ sites/geohub/src/components/pages/home/MapStyleCardList.svelte
