+++ sites/geohub/src/routes/(app)/+page.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
 	import { browser } from '$app/environment';
 	import { invalidate } from '$app/navigation';
-	import MapHero from '$components/MapHero.svelte';
+	import MapHero from '$components/pages/home/MapHero.svelte';
 	import DataUploadButton from '$components/pages/data/ingesting/DataUploadButton.svelte';
-	import MapStyleCardList from '$components/maps/MapStyleCardList.svelte';
+	import MapStyleCardList from '$components/pages/home/MapStyleCardList.svelte';
 	import { FooterItems, HeaderItems, MapStyleId } from '$lib/config/AppConfig';
 	import type { MapsData } from '$lib/types';
 	import {
